Fix: Create product
