+++ client/store/singleProduct.js
@@ -1,24 +1,12 @@
 import axios from 'axios'
+import {getAllProducts} from './products'
 
 const GET_SINGLE_PRODUCT = 'GET_SINGLE_PRODUCT'
-const CREATE_PRODUCT = 'CREATE_PRODUCT'
 
 export const getSingleProduct = product => ({
   type: GET_SINGLE_PRODUCT,
   product
 })
-
-export const _createProduct = product => ({
-  type: CREATE_PRODUCT,
-  product
-})
-export const createProduct = (product, history) => {
-  return async dispatch => {
-    const create = (await axios.post('/api/products/', product)).data
-    dispatch(_createProduct(create))
-    history.push('/home')
-  }
-}
 
 export const fetchSingleProduct = id => async dispatch => {
   try {
@@ -35,8 +23,6 @@ const initialState = []
 export default function singleProductReducer(state = initialState, action) {
   switch (action.type) {
     case GET_SINGLE_PRODUCT:
-      return action.product
-    case CREATE_PRODUCT:
       return action.product
     default:
       return state
